PodsDFVReactComponent: Encapsulate DFV-wide propTypes
