+++ ui/js/pods-dfv/_src/text/text.js
@@ -1,24 +1,12 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import { PodsDFVReactComponent } from 'pods-dfv/_src/core/react/pods-dfv-react-component';
 
-export class PodsDFVText extends React.Component {
+
+export class PodsDFVText extends PodsDFVReactComponent {
 	constructor ( props ) {
 		super( props );
 		this.state = {
 			value: this.props.fieldItemData[ 0 ]
-		};
-	}
-
-	static get propTypes () {
-		return {
-			fieldType: PropTypes.string,
-			fieldConfig: {
-				text_max_length: PropTypes.string,
-				text_placeholder: PropTypes.string,
-				readonly: PropTypes.bool
-			},
-			htmlAttr: PropTypes.object,
-			fieldItemData: PropTypes.array
 		};
 	}
